refactor: simplify database handling in AttributeStore and Graph initialization
