+++ graph/src/graph/graph.rs
@@ -73,10 +73,8 @@ use std::{
 };
 
 use atomic_refcell::AtomicRefCell;
-use fjall::Database;
 use itertools::Itertools;
 use lru::LruCache;
-use once_cell::sync::OnceCell;
 use orx_tree::DynTree;
 use parking_lot::Mutex;
 use roaring::RoaringTreemap;
@@ -397,22 +395,6 @@ fn drop_index_bg(
     );
 }
 
-static DATABASE: OnceCell<Database> = OnceCell::new();
-
-/// Get or initialize the shared fjall database for attribute stores.
-pub fn get_database() -> Database {
-    DATABASE
-        .get_or_init(|| {
-            Database::builder(format!("./attrs/{}", std::process::id()))
-                .temporary(true)
-                .manual_journal_persist(true)
-                .cache_size(128 * 1_024 * 1_024)
-                .open()
-                .expect("failed to open fjall database")
-        })
-        .clone()
-}
-
 impl Graph {
     #[must_use]
     pub fn new(
@@ -422,7 +404,6 @@ impl Graph {
         version: u64,
         name: &str,
     ) -> Self {
-        let db = get_database();
         Self {
             name: name.to_string(),
             node_cap: n,
@@ -440,8 +421,8 @@ impl Graph {
             all_nodes_matrix: VersionedMatrix::new(n, n),
             labels_matices: Vec::new(),
             relationship_matrices: Vec::new(),
-            node_attrs: AttributeStore::new(db.clone(), &format!("{name}/nodes"), version),
-            relationship_attrs: AttributeStore::new(db, &format!("{name}/relationships"), version),
+            node_attrs: AttributeStore::new(&format!("{name}/nodes"), version),
+            relationship_attrs: AttributeStore::new(&format!("{name}/relationships"), version),
             node_indexer: Indexer::default(),
             node_labels: Vec::new(),
             relationship_types: Vec::new(),
